Demo how to cache data on logging in
The list of nationalities is cached in the react router component
and passed as properties to the Profiles component. The same could
be done for other components which need the nationalities and
also the list of languages, avoiding them all loading these constant
values and storing them as state in each component.