Setup client to use tailwind CSS
The aim is to replace most (or all of the use of Semantic UI) with
Tailwind, avoiding the heavyweight components and opaque CSS that
are used everywhere and replacing them with plain HTML, styled with
TW classes.

Some specialized components may be required (datepickers, multiselects)
but these can either remain as they are or be replaced with other
options.