Fix build to work properly with Tailwind and PostCSS
The production build was broken because index.css wasn't created
by `npm build`. We use the instructions for setting up Tailwind
to work with React.

Uses craco in place of react-scripts to allow overriding of PostCSS
configuration. Tailwind needs PostCSS 8 which the current react-scripts
doesn't yet support so the PostCSS 7 compatibilty version has been
installed for now. This should be updated when react is compatible.

Packages such as Tailwinds apparently need to be in the 'dependencies'
section in package.json as the production client is build by the
heroku-postbuild command which ignores dev dependencies.